requirements.txt
flask==1.1.2
gunicorn==20.1.0
mysql-connector-python==8.0.24
requests==2.25.1
salt-pepper==0.7.6
psycopg2
susemanager_client
jira_client
